Project 3/requirements.txt
langchain==0.1.13
streamlit==1.32.2
openai==1.14.2
tiktoken==0.5.2
python-dotenv==1.0.1
faiss-cpu==1.8.0
langchain-openai==0.1.0



